EasyOCR/requirements.txt
easyocr
flask
werkzeug
pillow